+++ ichnaea/models/wifi.py
@@ -1,20 +1,34 @@
 import colander
+from enum import IntEnum
 from sqlalchemy import (
     Column,
+    Date,
     Index,
     String,
+    PrimaryKeyConstraint,
     UniqueConstraint,
 )
+from sqlalchemy.dialects.mysql import (
+    INTEGER as Integer,
+    TINYINT as TinyInteger,
+)
+from sqlalchemy.ext.declarative import declared_attr
 
+from ichnaea.models import constants
 from ichnaea.models.base import (
     _Model,
     BigIdMixin,
     CreationMixin,
+    PositionMixin,
+    TimeTrackingMixin,
 )
-from ichnaea.models import constants
 from ichnaea.models.hashkey import (
     HashKey,
     HashKeyQueryMixin,
+)
+from ichnaea.models.sa_types import (
+    MacColumn,
+    TinyIntEnum,
 )
 from ichnaea.models.schema import (
     CopyingSchema,
@@ -23,10 +37,13 @@ from ichnaea.models.schema import (
     ValidatorNode,
 )
 from ichnaea.models.station import (
+    BboxMixin,
     StationMixin,
     StationBlocklistMixin,
     ValidStationSchema,
 )
+
+WIFI_SHARDS = {}
 
 
 class WifiKey(HashKey):
@@ -145,3 +162,178 @@ class WifiBlocklist(WifiMixin, StationBlocklistMixin, _Model):
     _indices = (
         UniqueConstraint('key', name='wifi_blacklist_key_unique'),
     )
+
+
+class StationSource(IntEnum):
+    """
+    The station source states on what kind of data the station
+    record is based on. A lower integer value hints at a better quality
+    of the observation data that went into this station record.
+    """
+
+    fixed = 0  #: Outside knowledge about the true position of the station.
+    gnss = 3  #: Global navigation satellite system based data.
+    fused = 6  #: Observation data positioned based on fused data.
+    query = 9  #: Position estimate based on query data.
+
+
+class WifiMac(HashKey):
+
+    _fields = ('mac', )
+
+
+class WifiShard(HashKeyQueryMixin,
+                CreationMixin,
+                PositionMixin,
+                BboxMixin,
+                TimeTrackingMixin):
+
+    _hashkey_cls = WifiMac
+    _query_batch = 100
+
+    mac = Column(MacColumn(6))
+    radius = Column(Integer(unsigned=True))
+
+    country = Column(String(2))
+    samples = Column(Integer(unsigned=True))
+    source = Column(TinyIntEnum(StationSource))
+
+    block_first = Column(Date)
+    block_last = Column(Date)
+    block_count = Column(TinyInteger(unsigned=True))
+
+    @declared_attr
+    def __table_args__(cls):  # NOQA
+        _indices = (
+            PrimaryKeyConstraint('mac'),
+            Index('%s_country_idx' % cls.__tablename__, 'country'),
+            Index('%s_created_idx' % cls.__tablename__, 'created'),
+            Index('%s_modified_idx' % cls.__tablename__, 'modified'),
+            Index('%s_latlon_idx' % cls.__tablename__, 'lat', 'lon'),
+        )
+        return _indices + (cls._settings, )
+
+    @property
+    def range(self):
+        # BBB: alias
+        return self.radius
+
+    @property
+    def total_measures(self):
+        # BBB: alias
+        return self.samples
+
+    @classmethod
+    def shard_model(cls, mac):
+        """
+        Given a BSSID/MAC return the correct DB model class for this
+        shard of data.
+
+        The shard id is based on the fifth hex character of the vendor
+        prefix of the BSSID. This tends to be evenly distributed, but
+        still keeps data from the same vendor inside the same table.
+
+        It also allows us to later extend the sharding by taking in
+        parts of the sixth hex char without having to do a complete
+        re-sharding of everything, but merely breaking up each shard
+        further.
+        """
+        if not mac:
+            return None
+        return WIFI_SHARDS.get(mac.lower()[4], None)
+
+
+class WifiShard0(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_0'
+
+WIFI_SHARDS['0'] = WifiShard0
+
+
+class WifiShard1(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_1'
+
+WIFI_SHARDS['1'] = WifiShard1
+
+
+class WifiShard2(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_2'
+
+WIFI_SHARDS['2'] = WifiShard2
+
+
+class WifiShard3(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_3'
+
+WIFI_SHARDS['3'] = WifiShard3
+
+
+class WifiShard4(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_4'
+
+WIFI_SHARDS['4'] = WifiShard4
+
+
+class WifiShard5(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_5'
+
+WIFI_SHARDS['5'] = WifiShard5
+
+
+class WifiShard6(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_6'
+
+WIFI_SHARDS['6'] = WifiShard6
+
+
+class WifiShard7(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_7'
+
+WIFI_SHARDS['7'] = WifiShard7
+
+
+class WifiShard8(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_8'
+
+WIFI_SHARDS['8'] = WifiShard8
+
+
+class WifiShard9(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_9'
+
+WIFI_SHARDS['9'] = WifiShard9
+
+
+class WifiShardA(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_a'
+
+WIFI_SHARDS['a'] = WifiShardA
+
+
+class WifiShardB(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_b'
+
+WIFI_SHARDS['b'] = WifiShardB
+
+
+class WifiShardC(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_c'
+
+WIFI_SHARDS['c'] = WifiShardC
+
+
+class WifiShardD(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_d'
+
+WIFI_SHARDS['d'] = WifiShardD
+
+
+class WifiShardE(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_e'
+
+WIFI_SHARDS['e'] = WifiShardE
+
+
+class WifiShardF(WifiShard, _Model):
+    __tablename__ = 'wifi_shard_f'
+
+WIFI_SHARDS['f'] = WifiShardF